feature[test]: refactoring issue #510
